pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "brainmass-v3"
version = "0.1.0"
description = "Brainmass — an enterprise-grade agentic coding platform built on AWS Strands Agents SDK and Amazon Bedrock AgentCore"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    "strands-agents>=1.27.0",
    "strands-agents-tools>=0.2.20",
    "bedrock-agentcore>=1.3.0",
    "boto3>=1.34",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "hypothesis>=6.100",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "property: Property-based tests (Hypothesis)",
    "integration: Integration tests",
    "slow: Slow-running tests",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
